// src/types/enums/user/error-messages.enum.ts
export enum ErrorMessages {
  EXPIRED_TOKEN = "Le jeton JWT a expiré",
  INVALID_TOKEN = "Le jeton JWT est invalide",
  INVALID_OTP = "Le code OTP saisi est invalide ou a expiré.",
  OTP_ALREADY_USED = "Le code OTP saisi a déjà été utilisé.",
  USER_NOT_FOUND = "L'utilisateur n'existe pas.",
  CAR_NOT_FOUND = "La voiture n'existe pas.",
  ERROR_IN_ADDING_FCM = "Une erreur inattendue est survenue lors de l'ajout du token FCM, veuillez réessayer.",
  ERROR_IN_UPLOADING_IMAGE = "Une erreur inattendue est survenue lors de l'upload de l'image vers S3, veuillez réessayer.",
  ERROR_IN_LOGOUT = "Une erreur inattendue est survenue lors de la déconnexion, veuillez réessayer.",
  ERROR_IN_LISTING_MY_CARS = "Une erreur inattendue est survenue lors de la liste des voitures, veuillez réessayer.",
  ERROR_IN_REMOVING_FCM = "Une erreur inattendue est survenue lors de la suppression du token FCM, veuillez réessayer.",
  ERROR_IN_EDITING_CAR_DETAILS = "Une erreur inattendue est survenue lors de la mise à jour des détails de la voiture, veuillez réessayer.",
  ERROR_IN_GETTING_CAR_DETAILS = "Une erreur inattendue est survenue lors de la récupération des détails de la voiture, veuillez réessayer.",
  CANNOT_CREATE_USER_AT_THIS_MOMENT = "Une erreur est survenue lors de la création de l'utilisateur, veuillez réessayer.",
  CANNOT_RESET_PASSWORD_AT_THIS_MOMENT = "Une erreur est survenue lors de la réinitialisation du mot de passe, veuillez réessayer.",
  EMAIL_ALREADY_REGISTERED = "Un compte avec cet e-mail existe déjà.",
  OTP_ATTEMPTS_EXCEEDED = "Trop de tentatives OTP incorrectes. Veuillez réessayer plus tard.",
  UNAUTHORIZED_ACCESS = "Vous n'avez pas la permission d'accéder à cette ressource.",
  INVALID_CREDENTIALS = "Identifiants de connexion invalides",
  INVALID_OTP_REASON = "Raison OTP invalide. Veuillez réessayer avec la raison correcte.",
  ERROR_REGISTERING_CAR = "Une erreur est survenue lors de l'enregistrement de la voiture, veuillez réessayer.",
  INVALID_PASSWORD = "Le nouveau mot de passe ne peut pas être identique à l'ancien.",
  OLD_PASSWORD_DOESNOT_MATCH = "Ancien mot de passe invalide",
  ERROR_IN_CHANGE_PASSWORD = "Une erreur est survenue lors du changement de mot de passe, veuillez réessayer.",
  DATE_TIME_ERROR = "Date ou heure de début/fin invalide",
  CAR_ALREADY_BOOKED = "Cette voiture est déjà réservée par quelqu'un d'autre pendant ces dates",
  INVALID_BOOKING_SPAN = "La réservation doit durer au moins 2 heures",
  CANNOT_BOOK_SELF_OWNED_CAR = "Vous ne pouvez pas réserver votre propre voiture",
  CAR_NOT_AVAILABLE = "La voiture n'est pas disponible pendant ces dates",
  BOOKING_DOESNOT_EXIST = "La réservation n'existe pas",
  ERROR_IN_CANCELLING_BOOKING = "Vous ne pouvez annuler que les réservations à venir",
  TWENTY_FOUR_HOURS_PASSED = "Vous ne pouvez annuler une réservation que 24 heures avant l'heure de réservation",
  CHAT_NOT_FOUND = "La discussion entre ces utilisateurs n'existe pas",
  INCOMPLETE_HOST_PROFILE = "Veuillez compléter votre profil avant d'enregistrer une voiture",
  INVALID_CHAT_STATE = "La discussion n'a pas été correctement initialisée entre les deux utilisateurs",
  INVALID_START_END_DATE_TIME = "La date et l'heure de début doivent être inférieures à celles de fin",
}